feat: add redux-toolkit & move app in App.tsx
