+++ src/components/App.css
@@ -1,0 +1,33 @@
+@import-normalize;
+
+:root {
+  --white: #fffffe;
+  --medium-white: #f9f4ef;
+  --light-brown: #eaddcf;
+  --medium-brown: #8c7851;
+  --dark-brown: #716040;
+  --pink: #f25042;
+  --black: #020826;
+}
+
+html {
+  scroll-behavior: smooth;
+}
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
+body {
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen', 'Ubuntu',
+    'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue', sans-serif;
+  color: var(--dark-brown);
+}
+
+.container {
+  max-width: 95vw;
+  margin: 0 auto;
+}
