+++ src/components/App.tsx
@@ -1,0 +1,26 @@
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
+import { Provider } from 'react-redux';
+import store from './store';
+
+import './App.css';
+
+import Layout from './pages/Layout/Layout';
+import NoPage from './pages/NoPage/NoPage';
+import Home from './pages/Home/Home';
+
+function App() {
+  return (
+    <Provider store={store}>
+      <BrowserRouter>
+        <Routes>
+          <Route path='/' element={<Layout />}>
+            <Route index element={<Home />} />
+            <Route path='*' element={<NoPage />} />
+          </Route>
+        </Routes>
+      </BrowserRouter>
+    </Provider>
+  );
+}
+
+export default App;
