+++ src/components/store/cartSlice.ts
@@ -1,0 +1,11 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const cartSlice = createSlice({
+  name: 'cart',
+  initialState: {
+    products: [],
+  },
+  reducers: {},
+});
+
+export default cartSlice.reducer;
