+++ src/index.tsx
@@ -1,24 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import { BrowserRouter, Routes, Route } from 'react-router-dom';
-
-import './index.css';
-
-import Layout from './pages/Layout/Layout';
-import NoPage from './pages/NoPage/NoPage';
-import Home from './pages/Home/Home';
+import App from './components/App';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 
 root.render(
   <React.StrictMode>
-    <BrowserRouter>
-      <Routes>
-        <Route path='/' element={<Layout />}>
-          <Route index element={<Home />} />
-          <Route path='*' element={<NoPage />} />
-        </Route>
-      </Routes>
-    </BrowserRouter>
+    <App />
   </React.StrictMode>,
 );
